refactor: split bridge.go into domain-specific service files (Closes #324)
